refactor(wave7b-step3): make loader facade testless and thin
